+++ lib/cadmus/partial.rb
@@ -1,13 +1,13 @@
 module Cadmus
-  module Layout
+  module Partial
     extend ActiveSupport::Concern
     include Cadmus::Concerns::LiquidTemplateField
     include Cadmus::Concerns::ModelWithParent
 
     module ClassMethods
-      # Sets up a model to behave as a Cadmus layout.  This will add the following behaviors:
+      # Sets up a model to behave as a Cadmus partial.  This will add the following behaviors:
       #
-      # * A name field that determines the name of the layout for administrative UI
+      # * A name field that determines the name of the partial for administrative UI
       # * An optional, polymorphic +parent+ field
       # * A scope called +global+ that returns instances of this class that have no parent
       # * A +liquid_template+ method that parses the value of this model's +content+ field as a Liquid
@@ -16,7 +16,7 @@ module Cadmus
       #
       # @param options [Hash] options to modify the default behavior
       # @option options :name_field the name of the field to be used as the layout name.  Defaults to +:name+.
-      def cadmus_layout(options = {})
+      def cadmus_partial(options = {})
         model_with_parent
 
         cattr_accessor :name_field